+++ Pandora/Paradigm/Controlflow/Joint/Schemes/UTU.hs
@@ -1,7 +1,7 @@
-module Pandora.Pattern.Junction.Schemes.UTU (UTU (..)) where
+module Pandora.Paradigm.Controlflow.Joint.Schemes.UTU (UTU (..)) where
 
 import Pandora.Core.Functor (type (:.), type (>))
-import Pandora.Pattern.Junction.Interpreted (Interpreted (Primary, unwrap))
+import Pandora.Paradigm.Controlflow.Joint.Interpreted (Interpreted (Primary, unwrap))
 
 newtype UTU ct cu t u a = UTU (u :. t u > a)