+++ Pandora/Pattern.hs
@@ -1,5 +1,5 @@
 module Pandora.Pattern (module Exports) where
 
 import Pandora.Pattern.Object as Exports
-import Pandora.Pattern.Junction as Exports
+import Pandora.Paradigm.Controlflow.Joint as Exports
 import Pandora.Pattern.Functor as Exports