+++ Pandora/Pattern/Junction.hs
@@ -1,5 +1,0 @@
-module Pandora.Pattern.Junction (module Exports) where
-
-import Pandora.Pattern.Junction.Schemes as Exports
-import Pandora.Pattern.Junction.Transformer as Exports
-import Pandora.Pattern.Junction.Interpreted as Exports